move cube representations to own module
